requirements.txt
fastapi==0.115.4
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.9
